# Conflicts:
#	platform/lang-impl/src/com/intellij/ide/util/scopeChooser/ScopeChooserConfigurable.java

+++ java/java-impl/src/com/intellij/codeInsight/intention/impl/AddOnDemandStaticImportAction.java
@@ -1,18 +1,45 @@
-// Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.intellij.codeInsight.intention.impl;
 
 import com.intellij.codeInsight.highlighting.HighlightManager;
-import com.intellij.codeInsight.intention.BaseElementAtCaretIntentionAction;
 import com.intellij.java.JavaBundle;
+import com.intellij.modcommand.ActionContext;
+import com.intellij.modcommand.ModPsiUpdater;
+import com.intellij.modcommand.Presentation;
+import com.intellij.modcommand.PsiUpdateModCommandAction;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.colors.EditorColors;
-import com.intellij.openapi.editor.impl.ImaginaryEditor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Ref;
 import com.intellij.openapi.wm.WindowManager;
-import com.intellij.psi.*;
+import com.intellij.pom.java.JavaFeature;
+import com.intellij.psi.JavaPsiFacade;
+import com.intellij.psi.JavaRecursiveElementWalkingVisitor;
+import com.intellij.psi.JavaResolveResult;
+import com.intellij.psi.JavaTokenType;
+import com.intellij.psi.PsiCaseLabelElementList;
+import com.intellij.psi.PsiClass;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiErrorElement;
+import com.intellij.psi.PsiFile;
+import com.intellij.psi.PsiIdentifier;
+import com.intellij.psi.PsiImportList;
+import com.intellij.psi.PsiImportStatementBase;
+import com.intellij.psi.PsiImportStaticStatement;
+import com.intellij.psi.PsiJavaCodeReferenceElement;
+import com.intellij.psi.PsiJavaFile;
+import com.intellij.psi.PsiJavaToken;
+import com.intellij.psi.PsiManager;
+import com.intellij.psi.PsiMember;
+import com.intellij.psi.PsiMethod;
+import com.intellij.psi.PsiMethodCallExpression;
+import com.intellij.psi.PsiMethodReferenceExpression;
+import com.intellij.psi.PsiModifier;
+import com.intellij.psi.PsiReferenceParameterList;
+import com.intellij.psi.ResolveResult;
+import com.intellij.psi.codeStyle.JavaCodeStyleManager;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.psi.util.PsiUtil;
 import com.intellij.psi.util.PsiUtilCore;
@@ -20,22 +47,24 @@ import com.intellij.refactoring.RefactoringBundle;
 import com.intellij.util.IncorrectOperationException;
 import com.siyeh.ig.psiutils.ClassUtils;
 import com.siyeh.ig.psiutils.CommentTracker;
+import com.siyeh.ig.psiutils.ImportUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
-/**
- * @author ven
- */
-public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAction {
+public final class AddOnDemandStaticImportAction extends PsiUpdateModCommandAction<PsiIdentifier> {
   private static final Logger LOG = Logger.getInstance(AddOnDemandStaticImportAction.class);
+  
+  public AddOnDemandStaticImportAction() {
+    super(PsiIdentifier.class);
+  }
 
   @Override
-  @NotNull
-  public String getFamilyName() {
+  public @NotNull String getFamilyName() {
     return JavaBundle.message("intention.add.on.demand.static.import.family");
   }
 
@@ -45,29 +74,27 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
    * @param element     element to check
    * @return            target class that may be statically imported if any; {@code null} otherwise
    */
-  @Nullable
-  public static PsiClass getClassToPerformStaticImport(@NotNull PsiElement element) {
-    if (!PsiUtil.isLanguageLevel5OrHigher(element)) return null;
-    if (!(element instanceof PsiIdentifier) || !(element.getParent() instanceof PsiJavaCodeReferenceElement)) {
+  public static @Nullable PsiClass getClassToPerformStaticImport(@NotNull PsiElement element) {
+    if (!PsiUtil.isAvailable(JavaFeature.STATIC_IMPORTS, element)) return null;
+    if (!(element instanceof PsiIdentifier) || !(element.getParent() instanceof PsiJavaCodeReferenceElement refExpr)) {
       return null;
     }
     if (PsiTreeUtil.getParentOfType(element, PsiErrorElement.class, PsiImportStatementBase.class) != null) return null;
-    PsiJavaCodeReferenceElement refExpr = (PsiJavaCodeReferenceElement)element.getParent();
-    if (refExpr instanceof  PsiMethodReferenceExpression) return null;
+    if (refExpr instanceof PsiMethodReferenceExpression) return null;
     final PsiElement gParent = refExpr.getParent();
     if (gParent instanceof PsiMethodReferenceExpression) return null;
-    if (!(gParent instanceof PsiJavaCodeReferenceElement) ||
-        isParameterizedReference((PsiJavaCodeReferenceElement)gParent)) return null;
+    if (!(gParent instanceof PsiJavaCodeReferenceElement parentRef)) return null;
+    if (isParameterizedReference(parentRef)) return null;
 
-    if (PsiUtilCore.getElementType(PsiTreeUtil.nextCodeLeaf(gParent)) == JavaTokenType.ARROW) {
-      return null;
+    if (PsiUtilCore.getElementType(PsiTreeUtil.nextCodeLeaf(gParent)) == JavaTokenType.ARROW &&
+        !(gParent.getParent() instanceof PsiCaseLabelElementList)) {
+        return null;
     }
 
     PsiElement resolved = refExpr.resolve();
-    if (!(resolved instanceof PsiClass)) {
+    if (!(resolved instanceof PsiClass psiClass)) {
       return null;
     }
-    PsiClass psiClass = (PsiClass)resolved;
     if (PsiUtil.isFromDefaultPackage(psiClass) ||
         psiClass.hasModifierProperty(PsiModifier.PRIVATE) ||
         psiClass.getQualifiedName() == null) return null;
@@ -78,11 +105,17 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
       final PsiElement qualifier = call.getMethodExpression().getQualifier();
       if (qualifier == null) return null;
       qualifier.delete();
-      final PsiMethod method = call.resolveMethod();
-      if (method != null && method.getContainingClass() != psiClass)  return null;
+      JavaResolveResult[] results = call.multiResolve(false);
+      if (Arrays.stream(results)
+        .map(ResolveResult::getElement)
+        .anyMatch(psiElement ->
+                    psiElement instanceof PsiMethod psiMethod &&
+                    psiMethod.getContainingClass() != psiClass)) {
+        return null;
+      }
     }
     else {
-      PsiElement refNameElement = ((PsiJavaCodeReferenceElement)gParent).getReferenceNameElement();
+      PsiElement refNameElement = parentRef.getReferenceNameElement();
       if (refNameElement == null) return null;
       final PsiJavaCodeReferenceElement copy = JavaPsiFacade.getElementFactory(refNameElement.getProject())
         .createReferenceFromText(refNameElement.getText(), refExpr);
@@ -100,29 +133,45 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
           }
         }
       }
-      PsiElement resolve = ((PsiJavaCodeReferenceElement)gParent).resolve();
-      if (resolve instanceof PsiMember && !((PsiMember)resolve).hasModifierProperty(PsiModifier.STATIC)) return null;
+      if (parentRef.resolve() instanceof PsiMember member && !member.hasModifierProperty(PsiModifier.STATIC)) return null;
     }
 
-    PsiFile file = refExpr.getContainingFile();
-    if (!(file instanceof PsiJavaFile)) return null;
-    PsiImportList importList = ((PsiJavaFile)file).getImportList();
-    if (importList == null) return null;
-
-    return psiClass;
+    if (refExpr.getContainingFile() instanceof PsiJavaFile javaFile && javaFile.getImportList() != null) {
+      return psiClass;
+    }
+    return null;
   }
 
   @Override
-  public boolean isAvailable(@NotNull Project project, Editor editor, @NotNull PsiElement element) {
+  protected @Nullable Presentation getPresentation(@NotNull ActionContext context, @NotNull PsiIdentifier element) {
     PsiClass classToImport = getClassToPerformStaticImport(element);
-    if (classToImport != null) {
-      String text = JavaBundle.message("intention.add.on.demand.static.import.text", classToImport.getQualifiedName());
-      setText(text);
+    if (classToImport == null) {
+      return null;
     }
-    return classToImport != null;
+    Project project = element.getProject();
+    JavaCodeStyleManager codeStyleManager = JavaCodeStyleManager.getInstance(project);
+    if (codeStyleManager.isStaticAutoImportName(classToImport.getQualifiedName())) {
+      return null;
+    }
+    return Presentation.of(JavaBundle.message("intention.add.on.demand.static.import.text", classToImport.getQualifiedName()));
   }
 
-  public static boolean invoke(final Project project, PsiFile file, final Editor editor, PsiElement element) {
+  public static boolean invoke(final Project project, PsiFile file, final Editor editor, @NotNull PsiElement element) {
+    List<PsiJavaCodeReferenceElement> dequalifiedElements = new ArrayList<>();
+    boolean conflicts = addStaticImports(file, element, dequalifiedElements);
+    if (editor != null) {
+      ApplicationManager.getApplication().invokeLater(() -> {
+        if (collectChangedPlaces(project, editor, dequalifiedElements)) {
+          WindowManager.getInstance().getStatusBar(project).setInfo(RefactoringBundle.message("press.escape.to.remove.the.highlighting"));
+        }
+      }, project.getDisposed());
+    }
+    return conflicts;
+  }
+
+  private static boolean addStaticImports(@NotNull PsiFile file,
+                                          @NotNull PsiElement element,
+                                          @NotNull List<@NotNull PsiJavaCodeReferenceElement> dequalifiedElements) {
     final PsiJavaCodeReferenceElement refExpr = (PsiJavaCodeReferenceElement)element.getParent();
     final PsiClass aClass = (PsiClass)refExpr.resolve();
     if (aClass == null) {
@@ -130,17 +179,24 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
     }
     final PsiClass containingClass = PsiUtil.getTopLevelClass(refExpr);
     if (aClass != containingClass || !ClassUtils.isInsideClassBody(element, aClass)) {
-      PsiImportList importList = ((PsiJavaFile)file).getImportList();
+      PsiJavaFile psiJavaFile = (PsiJavaFile)file;
+      PsiImportList importList = psiJavaFile.getImportList();
       if (importList == null) {
         return false;
       }
       boolean alreadyImported = false;
-      for (PsiImportStaticStatement statement : importList.getImportStaticStatements()) {
-        if (!statement.isOnDemand()) continue;
-        PsiClass staticResolve = statement.resolveTargetClass();
-        if (aClass == staticResolve) {
-          alreadyImported = true;
-          break;
+      String qualifiedName = aClass.getQualifiedName();
+      if (qualifiedName != null && ImportUtils.createImplicitImportChecker(psiJavaFile).isImplicitlyImported(qualifiedName + ".*", true)) {
+        alreadyImported = true;
+      }
+      if (!alreadyImported) {
+        for (PsiImportStaticStatement statement : importList.getImportStaticStatements()) {
+          if (!statement.isOnDemand()) continue;
+          PsiClass staticResolve = statement.resolveTargetClass();
+          if (aClass == staticResolve) {
+            alreadyImported = true;
+            break;
+          }
         }
       }
       if (!alreadyImported) {
@@ -161,22 +217,24 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
       copy.accept(new JavaRecursiveElementWalkingVisitor() {
         int delta;
         @Override
-        public void visitReferenceElement(PsiJavaCodeReferenceElement expression) {
-          if (isParameterizedReference(expression) ||
-              expression instanceof PsiMethodReferenceExpression ||
-              expression.getParent() instanceof PsiErrorElement) {
-            super.visitElement(expression);
-            return;
-          }
-          PsiElement qualifierExpression = expression.getQualifier();
-          if (qualifierExpression instanceof PsiJavaCodeReferenceElement && ((PsiJavaCodeReferenceElement)qualifierExpression).isReferenceTo(aClass)) {
+        public void visitReferenceElement(@NotNull PsiJavaCodeReferenceElement expression) {
+          PsiElement qualifierExpression;
+          if (!isParameterizedReference(expression) &&
+              !(expression instanceof PsiMethodReferenceExpression) &&
+              !(expression.getParent() instanceof PsiErrorElement) &&
+              (qualifierExpression = expression.getQualifier()) instanceof PsiJavaCodeReferenceElement &&
+              ((PsiJavaCodeReferenceElement)qualifierExpression).isReferenceTo(aClass)) {
             try {
-              PsiElement resolved = expression.resolve();
+              JavaResolveResult[] resolved = expression.multiResolve(false);
               int end = expression.getTextRange().getEndOffset();
               qualifierExpression.delete();
+              PsiElement firstChild = expression.getFirstChild();
+              if (firstChild instanceof PsiJavaToken && ((PsiJavaToken)firstChild).getTokenType() == JavaTokenType.DOT) {
+                firstChild.delete();
+              }
               delta += end - expression.getTextRange().getEndOffset();
-              PsiElement after = expression.resolve();
-              if (manager.areElementsEquivalent(after, resolved)) {
+              JavaResolveResult[] resolvedAfter = expression.multiResolve(false);
+              if (resolvesToSame(manager, resolved, resolvedAfter)) {
                 int offset = expression.getTextRange().getStartOffset() + delta;
                 PsiJavaCodeReferenceElement originalExpression =
                   PsiTreeUtil.findElementOfClassAtOffset(root, offset, PsiJavaCodeReferenceElement.class, false);
@@ -196,25 +254,32 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
               LOG.error(e);
             }
           }
-          super.visitElement(expression);
+          super.visitReferenceElement(expression);
         }
       });
 
       for (PsiJavaCodeReferenceElement expression : expressionsToDequalify) {
         new CommentTracker().deleteAndRestoreComments(Objects.requireNonNull(expression.getQualifier()));
       }
-      if (editor != null && !(editor instanceof ImaginaryEditor)) {
-        ApplicationManager.getApplication().invokeLater(() -> {
-          if (collectChangedPlaces(project, editor, expressionsToDequalify)) {
-            WindowManager.getInstance().getStatusBar(project).setInfo(RefactoringBundle.message("press.escape.to.remove.the.highlighting"));
-          }
-        }, project.getDisposed());
-      }
+      dequalifiedElements.addAll(expressionsToDequalify);
     }
     return conflict.get();
   }
 
-  private static boolean collectChangedPlaces(Project project, Editor editor, List<PsiJavaCodeReferenceElement> expressionsToDequalify) {
+  private static boolean resolvesToSame(@NotNull PsiManager manager, JavaResolveResult @NotNull [] resolved, JavaResolveResult @NotNull [] resolvedAfter) {
+    // returns true if there's at least one element from "resolved" which is the same as one of "resolvedAfter"
+    for (JavaResolveResult result : resolved) {
+      if (!result.isAccessible()) continue;
+      for (JavaResolveResult resultAfter : resolvedAfter) {
+        if (resultAfter.isAccessible() && manager.areElementsEquivalent(result.getElement(), resultAfter.getElement())) {
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+
+  private static boolean collectChangedPlaces(Project project, Editor editor, @NotNull List<? extends PsiJavaCodeReferenceElement> expressionsToDequalify) {
     boolean found = false;
     for (PsiJavaCodeReferenceElement expression : expressionsToDequalify) {
       if (!expression.isValid()) continue;
@@ -227,8 +292,12 @@ public class AddOnDemandStaticImportAction extends BaseElementAtCaretIntentionAc
   }
 
   @Override
-  public void invoke(@NotNull Project project, Editor editor, @NotNull PsiElement element) throws IncorrectOperationException {
-    invoke(project, element.getContainingFile(), editor, element);
+  protected void invoke(@NotNull ActionContext context, @NotNull PsiIdentifier element, @NotNull ModPsiUpdater updater) {
+    List<PsiJavaCodeReferenceElement> dequalifiedElements = new ArrayList<>();
+    addStaticImports(element.getContainingFile(), element, dequalifiedElements);
+    for (PsiJavaCodeReferenceElement ref : dequalifiedElements) {
+      updater.highlight(ref);
+    }
   }
 
   private static boolean isParameterizedReference(final PsiJavaCodeReferenceElement expression) {
